Add highlighting of section titles
A section starts with a number of consecutive `#` at the beginning of
the line, followed by its title on the same line. The number of `#` is
the level of the section, and each level gets a separate face.

At this point, maybe I should just transform the Discourse format into
markdown and use markdown-mode to highlight the buffer...