Improve parsing of user names
Allow dots in the middle of user names, as they do happen in the wild.

However, don't allow a dot at the end, as it leads to an ambiguity
when the user name is followed by a sentence-ending period. (This
means user names ending with a dot can't be parsed properly, but we're
resolving the ambiguity in favor of the vastly more common cases.)